test(events): split mouse event kind coverage
